Fix writing LAZ 1.4 files by piping through laszip
The problem was that the Las1.4 class overrides the write method because the 1.4 header needs additional work for the evlrs.

The problem was that we were writing the header as the very last step of the method just like its done in the base class. but when the data was piped to laszip
we were rewriting the header as if the data was not compressed

Kinda hard to explain