Add append mode (mode='a') to pylas.open
Append mode can be used to append points to an existing file
without having to load the file in memory.

Points will be written after the last point.

This mode works on LAS and LAZ files.
For LAZ files, only the lazrs backend can be used as using laszip.exe does not have all the
feature required to be able to implement an append mode.