Make do_compress optional if a LazBackend is provided
If a laz_backend is provided to the LasWriter (or pylas.open)
then it is not mandatory to also provide the do_compress parameter.

If laz_backend is provided and do_compress is None -> data will
be compressed.

If laz_backend is provided and do_compress is False -> data will not
be compressed

small change to avoid redundancy